Channels (#3)
- InputManager changed to manage input.
- Created ChannelBase and ChannelListener classes.
- Created Channel and SubChannel classes.
- Mouse axis events are no longer separated into x and y axes and have Vector2f (i.e. MouseAxisEvent has information for both axes)
- Controller axis events are no longer separated into left_stick_x, left_stick_y, right_stick_x, etc. but are now events that have all the information for one stick (i.e. a ControllerAxisEvent has information for both left_stick_x and left_stick_y).
- Controller trigger events are now treated separately to stick events.

Each action is now represented by a Channel. Each Channel manages one or more SubChannels which correspond to a particular raw input. The SubChannels add observers to the event dispatcher so do not need to be updated in any way. Channels have listeners which they notify upon a change in the Channel value. 

The input manager has been changed so it actually manages the input as opposed to just containing all the input information as it was previously. I.e. InputManager keeps a reference to BaseGame and executes BaseGame::close() as opposed to BaseGame querying InputManager every frame and asking if it should close itself.